refactor(configurator): unify field render strategy with  helper
