+++ src/features/form-configurator/configuratorMap.ts
@@ -1,28 +1,59 @@
-import Configurator__Text from "./configurators/Configurator__Text";
-import Configurator__Email from "./configurators/Configurator__Email";
-import Configurator__Password from "./configurators/Configurator__Password";
-import Configurator__Number from "./configurators/Configurator__Number";
-import Configurator__Select from "./configurators/Configurator__Select";
-import Configurator__Checkbox from "./configurators/Configurator__Checkbox";
-import Configurator__Radio from "./configurators/Configurator__Radio";
+import React, { JSX } from "react";
+import LABEL from "./configurator-fields/LABEL";
+import NAME from "./configurator-fields/NAME";
+import PLACEHOLDER from "./configurator-fields/PLACEHOLDER";
+import OPTIONS from "./configurator-fields/OPTIONS";
+import { FieldRenderProps, FieldTypeSchema } from "./GenericConfigurator";
 
-type FieldType =
-  | "text"
-  | "email"
-  | "password"
-  | "number"
-  | "select"
-  | "checkbox"
-  | "radio";
+export const asField =
+  (
+    Component: React.ComponentType<any>
+  ): ((props: FieldRenderProps) => JSX.Element) =>
+  (props) =>
+    React.createElement(Component, props);
 
-const configuratorMap: Record<FieldType, React.ComponentType<any>> = {
-  text: Configurator__Text,
-  email: Configurator__Email,
-  password: Configurator__Password,
-  number: Configurator__Number,
-  select: Configurator__Select,
-  checkbox: Configurator__Checkbox,
-  radio: Configurator__Radio,
+export const inputTypeSchemas: Record<string, FieldTypeSchema> = {
+  text: {
+    fields: [asField(NAME), asField(LABEL), asField(PLACEHOLDER)],
+    validation: { required: "This field is required" },
+  },
+  email: {
+    fields: [asField(NAME), asField(LABEL), asField(PLACEHOLDER)],
+    validation: {
+      required: "Email is required",
+      pattern: {
+        value: /^[^\s@]+@[^\s@]+\.[^\s@]+$/,
+        message: "Invalid email format",
+      },
+    },
+  },
+  number: {
+    fields: [asField(NAME), asField(LABEL), asField(PLACEHOLDER)],
+    validation: {
+      required: "Age is required",
+      min: { value: 18, message: "You must be at least 18" },
+    },
+  },
+  password: {
+    fields: [asField(NAME), asField(LABEL), asField(PLACEHOLDER)],
+    validation: {
+      required: "Password is required",
+      minLength: { value: 6, message: "Minimum 6 characters" },
+    },
+  },
+  checkbox: {
+    fields: [asField(OPTIONS), asField(NAME), asField(LABEL)],
+    validation: { required: "Please select 1 option" },
+    optionKey: "checkbox-options",
+  },
+  radio: {
+    fields: [asField(OPTIONS), asField(NAME), asField(LABEL)],
+    validation: { required: "Please select at least 1 option" },
+    optionKey: "radio-options",
+  },
+  select: {
+    fields: [asField(OPTIONS), asField(NAME), asField(LABEL)],
+    validation: { required: "Please select at least 1 option" },
+    optionKey: "select-options",
+  },
 };
-
-export default configuratorMap;
